+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 -----
 
 - #99 Include out-of-stock analyses for export file
+- #103 Enable date range filtering using cookies instead of session storage
 - #98 Capture Tamanu ID in Tupaia export file
 - #95 Decouple Tamanu DiagnosticReport notifications from publication workflow
 - #92 Add the keyword next to the analysis name in stats report